+++ integration-test/entrypoint.clj
@@ -9,18 +9,16 @@
    [llm-mock.server :as llm-mock.server]))
 
 (def namespaces
-  '[
-    ;; integration.initialize-test
-    ;; integration.chat.hooks-test
-    ;; integration.chat.openai-test
-    ;; integration.chat.anthropic-test
-    ;; integration.chat.github-copilot-test
-    ;; integration.chat.google-test
-    ;; integration.chat.ollama-test
-    ;; integration.chat.custom-provider-test
+  '[integration.initialize-test
+    integration.chat.openai-test
+    integration.chat.anthropic-test
+    integration.chat.github-copilot-test
+    integration.chat.google-test
+    integration.chat.ollama-test
+    integration.chat.custom-provider-test
+    integration.chat.hooks-test
     integration.chat.commands-test
-    ;;integration.rewrite.openai-test
-    ])
+    integration.rewrite.openai-test])
 
 (defn timeout [timeout-ms callback]
   (let [fut (future (callback))
@@ -42,7 +40,6 @@
   `(binding [original-report t/report
              t/report log-tail-report]
      ~@body))
-
 
 (def tinyproxy-dir-env-var
   "Env var name pointing to the directory of the Tinyproxy executable."